Added corequisite handling
And yes, adding a capital C to every prerequisite that is a corequisite
was the best idea I had. Also designated AP Physics with corequisites to
test it (it worked). Also cleaned up an empty prereq set in Calculus AB
that was making it always show as working. That's worrying. I wonder if
there's any more like that.

Fixes #37.